Renamed every __init__ to main; creäte delegating __init__s
