+++ src/phonology/inventory.py
@@ -1,9 +1,9 @@
 
 """The phoneme inventory editor"""
 
-import phonology
+from . import main
 
-class InventoryView(phonology.PhonologyView):
+class InventoryView(main.PhonologyView):
     def build(self):
         # make a split screen with two consonant charts and two vowel trapeziums;
         # one on top for the phonology being edited, one at the bottom for complete IPA
